rename construct.py to fix unittest imports
